move generateAndWriteFileSummary to summaries
