+++ modules/summaries.js
@@ -114,11 +114,45 @@ async function getSummaries(dir, test){
   return summaries
 }
 
+
+/**
+ * Processes a file by generating a summary using the specified machine learning model
+ * and writing the summary to a new file.
+ * @param {string} dir - The directory of the file being processed.
+ * @param {string} filePathRelative - The relative path of the file being processed.
+ * @param {string} fileContent - The content of the file being processed.
+ * @param {object} model - The machine learning model used to generate the summary.
+ */
+async function generateAndWriteFileSummary(dir, filePathRelative, fileContent, model) {
+  const fileSummary = require('./agents/indexer');
+
+  try {
+    const output = await fileSummary(fileContent, model);
+
+    if (output) {
+      const filePathFull = path.join(dir, filePathRelative);
+      const summaryFilePath = path.join(filePathFull + '.ai.txt');
+      const summaryFileContent = `File Path: ${filePathRelative}\nSummary:\n${output}`;
+      fs.writeFileSync(summaryFilePath, summaryFileContent);
+      const timestamp = new Date().toISOString();
+      const hour = timestamp.match(/\d\d:\d\d/);
+
+      console.log(`${hour}: Updated ${summaryFilePath}`);
+    }
+  } catch (error) {
+    console.error(`Error processing file: ${filePathRelative}`, error);
+  }
+}
+exports.generateAndWriteFileSummary = generateAndWriteFileSummary;
+;
+
+
 module.exports = {
     readAllSummaries,
     getFiles,
     types,
     getSummaries,
     chunkSummaries,
-    maxSummaryTokenCount
+    maxSummaryTokenCount,
+    generateAndWriteFileSummary
 }
